Change the logic to avoid undesired tag removals
Before this commit, if the user forgets to save the manifest and then
tries to dump it, all the new tags in their nodes would be removed. Now
this effect, that also could be desired, is controlled and by default
only remove tags if the last manifest has been saved since other
machine.